User Acquisition Modal: Refactor to tsx and use Modal.tsx
