+++ static/css/timetable/partials/save_calendar_modal.scss
@@ -1,8 +1,4 @@
 .abnb-modal {
-  height: 465px;
-  max-width: 450px;
-  overflow: auto;
-
   > div {
     outline: none;
   }
