+++ static/css/timetable/partials/user_acquisition_modal.scss
@@ -1,8 +1,4 @@
 .user-acquisition-modal {
-  height: 495px;
-  max-width: 450px;
-  overflow: auto;
-
   .user-acquisition-modal__container {
     padding: 30px;
   }
