+++ static/js/redux/ui/Semesterly.tsx
@@ -28,7 +28,7 @@ import SignupModalContainer from "./containers/modals/signup_modal_container";
 import TutModalContainer from "./containers/modals/tut_modal_container";
 import PeerModalContainer from "./containers/modals/peer_modal_container";
 import SaveCalendarModalContainer from "./containers/modals/save_calendar_modal_container";
-import UserAcquisitionModal from "./modals/user_acquisition_modal";
+import UserAcquisitionModal from "./modals/UserAcquisitionModal";
 import TermsOfServiceModalContainer from "./containers/terms_of_service_modal_container";
 import TermsOfServiceBannerContainer from "./containers/terms_of_service_banner_container";
 import UserSettingsModal from "./modals/UserSettingsModal";
